+++ test/cldr_messages_test.exs
@@ -3,28 +3,28 @@ defmodule Cldr_Messages_Test do
   doctest Cldr.Message
 
   test "simple formatting of a number" do
-    assert Cldr.Message.to_string("You have {number} jelly beans", number: 1234) ==
-      {:ok, "You have 1,234 jelly beans"}
+    assert Cldr.Message.format("You have {number} jelly beans", number: 1234) ==
+             {:ok, "You have 1,234 jelly beans"}
 
-    assert Cldr.Message.to_string("You have {number} jelly beans", number: Decimal.new(1234)) ==
-      {:ok, "You have 1,234 jelly beans"}
+    assert Cldr.Message.format("You have {number} jelly beans", number: Decimal.new(1234)) ==
+             {:ok, "You have 1,234 jelly beans"}
   end
 
   test "formatting a unit" do
-    assert Cldr.Message.to_string("I am {height} tall", height: Cldr.Unit.new(1.8, :meter)) ==
-      {:ok, "I am 1.8 meters tall"}
+    assert Cldr.Message.format("I am {height} tall", height: Cldr.Unit.new(1.8, :meter)) ==
+             {:ok, "I am 1.8 meters tall"}
   end
 
   test "formatting money" do
-    assert Cldr.Message.to_string("I have {money}", money: Money.new(:USD, 123)) ==
-      {:ok, "I have $123.00"}
+    assert Cldr.Message.format("I have {money}", money: Money.new(:USD, 123)) ==
+             {:ok, "I have $123.00"}
 
-    assert Cldr.Message.to_string("I have {money, money, short}", money: Money.new(:USD, 123)) ==
-      {:ok, "I have $123"}
+    assert Cldr.Message.format("I have {money, money, short}", money: Money.new(:USD, 123)) ==
+             {:ok, "I have $123"}
   end
 
   test "arguments are inserted positionally" do
-    assert Cldr.Message.format("{0} {1, plural,
+    assert Cldr.Message.format_to_list("{0} {1, plural,
       one {est {2, select, female {allée} other  {allé}}}
       other {sont {2, select, female {allées} other {allés}}}
     } à {3}", ["You", 3, "female", "three"], []) ==
@@ -32,7 +32,7 @@ defmodule Cldr_Messages_Test do
   end
 
   test "keyword selectors" do
-    assert Cldr.Message.format("You have {item_count, plural,
+    assert Cldr.Message.format_to_list("You have {item_count, plural,
         =0 {no items}
         one {1 item}
         other {{item_count} items}
@@ -40,7 +40,7 @@ defmodule Cldr_Messages_Test do
   end
 
   test "numbers in percent format" do
-    assert Cldr.Message.format("{taxable_area, select,
+    assert Cldr.Message.format_to_list("{taxable_area, select,
         yes {An additional {tax_rate, number, percent} tax will be collected.}
         other {No taxes apply.}
     }", taxable_area: "yes", tax_rate: 0.15) ==
@@ -48,14 +48,14 @@ defmodule Cldr_Messages_Test do
   end
 
   test "select other category" do
-    assert Cldr.Message.format("{taxable_area, select,
+    assert Cldr.Message.format_to_list("{taxable_area, select,
         yes {An additional {tax_rate, number, percent} tax will be collected.}
         other {No taxes apply.}
     }", taxable_area: "no", tax_rate: 0.15) == {:ok, ["No taxes apply."]}
   end
 
   test "explicit selector" do
-    assert Cldr.Message.format("You have {item_count, plural,
+    assert Cldr.Message.format_to_list("You have {item_count, plural,
         =0 {no items}
         one {1 item}
         other {{item_count} items}
@@ -65,7 +65,7 @@ defmodule Cldr_Messages_Test do
   end
 
   test "other keyword selector" do
-    assert Cldr.Message.format("You have {item_count, plural,
+    assert Cldr.Message.format_to_list("You have {item_count, plural,
         =0 {no items}
         one {1 item}
         other {{item_count} items}
@@ -75,7 +75,7 @@ defmodule Cldr_Messages_Test do
   end
 
   test "nested complex arguments" do
-    assert Cldr.Message.format(
+    assert Cldr.Message.format_to_list(
              "{gender_of_host, select,
       female {
         {num_guests, plural, offset: 1
@@ -95,13 +95,16 @@ defmodule Cldr_Messages_Test do
           =1 {{host} invites {guest} to their party.}
           =2 {{host} invites {guest} and one other person to their party.}
           other {{host} invites {guest} and # other people to their party.}}}}",
-      gender_of_host: "male", host: "kip", guest: "jim", num_guests: 1
+             gender_of_host: "male",
+             host: "kip",
+             guest: "jim",
+             num_guests: 1
            ) ==
              {:ok, ["kip invites jim to his party."]}
   end
 
   test "offset" do
-    assert Cldr.Message.format(
+    assert Cldr.Message.format_to_list(
              "{gender_of_host, select,
       female {
         {num_guests, plural, offset: 1
@@ -121,13 +124,16 @@ defmodule Cldr_Messages_Test do
           =1 {{host} invites {guest} to their party.}
           =2 {{host} invites {guest} and one other person to their party.}
           other {{host} invites {guest} and # other people to their party.}}}}",
-      gender_of_host: "male", host: "kip", guest: "jim", num_guests: 4
+             gender_of_host: "male",
+             host: "kip",
+             guest: "jim",
+             num_guests: 4
            ) ==
              {:ok, ["kip invites jim and 3 other people to his party."]}
   end
 
   test "select ordinal" do
-    assert Cldr.Message.format("It's my cat's {year, selectordinal,
+    assert Cldr.Message.format_to_list("It's my cat's {year, selectordinal,
         one {#st}
         two {#nd}
         few {#rd}
@@ -138,27 +144,31 @@ defmodule Cldr_Messages_Test do
   end
 
   test "decimal selector" do
-    assert Cldr.Message.format("{a, plural, =0 {zero} other {other #}}", a: Decimal.new(2)) ==
+    assert Cldr.Message.format_to_list("{a, plural, =0 {zero} other {other #}}", a: Decimal.new(2)) ==
              {:ok, ["other 2"]}
 
-    assert Cldr.Message.format("{a, plural, =0 {zero} other {other #}}", a: Decimal.new(0)) ==
+    assert Cldr.Message.format_to_list("{a, plural, =0 {zero} other {other #}}", a: Decimal.new(0)) ==
              {:ok, ["zero"]}
   end
 
   test "plural, select and selectordinal with no other argument raises" do
-    assert Cldr.Message.format("{num, plural, =0 {it's zero}}", num: 1) ==
+    assert Cldr.Message.format_to_list("{num, plural, =0 {it's zero}}", num: 1) ==
              {:error,
               {Cldr.Message.ParseError,
                "'plural', 'select' and 'selectordinal' arguments must have an 'other' clause. Found %{0 => [literal: \"it's zero\"]}"}}
 
-    assert Cldr.Message.format("{num, select, zero {it's zero}}", num: 1) ==
+    assert Cldr.Message.format_to_list("{num, select, zero {it's zero}}", num: 1) ==
              {:error,
               {Cldr.Message.ParseError,
                "'plural', 'select' and 'selectordinal' arguments must have an 'other' clause. Found %{\"zero\" => [literal: \"it's zero\"]}"}}
 
-    assert Cldr.Message.format("{num, selectordinal, =0 {it's zero}}", num: 1) ==
+    assert Cldr.Message.format_to_list("{num, selectordinal, =0 {it's zero}}", num: 1) ==
              {:error,
               {Cldr.Message.ParseError,
                "'plural', 'select' and 'selectordinal' arguments must have an 'other' clause. Found %{0 => [literal: \"it's zero\"]}"}}
+  end
+
+  test "the format macro" do
+    assert MessageMacro.test() == "this is my message with a 1"
   end
 end