+++ CHANGELOG.md
@@ -1,3 +1,31 @@
+# Changelog for Cldr_Messages v0.3.0
+
+This is the changelog for Cldr_Messages v0.3.0 released on August 29th, 2019.  For older changelogs please consult the release tag on [GitHub](https://github.com/elixir-cldr/cldr_messages/tags)
+
+### Breaking Changes
+
+* Standardize on the `Cldr.Message.format/3` as the public api. `Cldr.Message.to_string/3` is removed.
+
+### Enhancements
+
+* Add the macro `<backend>.Cldr.Message.format/3` to parse messages at compile time as a way to optiise performance at runtime. To use it add `import <backend>.Cldr.Message` to your module and use `format/3`.  An example:
+
+```elixir
+defmodule SomeModule do
+  import MyApp.Cldr.Message
+
+  def my_function do
+    format("this is a string with a param {param}", param: 3)
+  end
+end
+```
+
+* Add `Cldr.Mesasge.format_to_list/3` formats to an `io_list`
+
+### Bug Fixes
+
+* Fix dialyzer warnings.  There are some warnings from combinators that will require `nimble_parsec` version 0.5.2 to be published before they are resolved.
+
 # Changelog for Cldr_Messages v0.2.0
 
 This is the changelog for Cldr_Messages v0.2.0 released on August 27th, 2019.  For older changelogs please consult the release tag on [GitHub](https://github.com/elixir-cldr/cldr_messages/tags)